+++ CHANGELOG.md
@@ -19,6 +19,7 @@ CHANGELOG
  * Fix log errors correctly and define user errors as notices
  * Resolve mutation payload class automatically for easy override
  * Add support tu use CRUD extensions for real PHP interfaces without register a graphql interface type.
+ * [BC BREAK] Removed `getPriority` method in CRUD extensions, must use service tag priority instead. 
 
 1.0.6 - 18-03-23
 ----